+++ src/components/MovieCard.jsx
@@ -9,19 +9,22 @@ const MovieCard = (props) => {
         <h3 className="card-title text-3xl text-center">
           {props.moviesObject.title}
         </h3>
-        <span className="badge mx-1 badge-secondary text-2xl p-4">
+        <span className="badge mx-1 badge-primary text-2xl p-4 w-auto h-auto">
           {props.moviesObject.year}
         </span>
-        <h4 className="badge mx-1 badge-primary text-2xl p-4">
+        <h4 className="badge mx-1 badge-secondary text-2xl p-4 w-auto h-auto">
           {props.moviesObject.director}
         </h4>
 
-        <p className="badge mx-1 badge-outline text-2xl p-4">
+        <p className="badge mx-1 badge-outline text-2xl p-4 w-auto h-auto">
           {props.moviesObject.duration}
         </p>
         <div className="flex flex-wrap items-center justify-center">
           {props.moviesObject.genre.map((oneGenre, index) => (
-            <span className="badge m-1 badge-accent text-2xl p-4" key={index}>
+            <span
+              className="badge m-1 badge-accent text-2xl p-4 w-auto h-auto"
+              key={index}
+            >
               {oneGenre}
             </span>
           ))}
