+++ src/components/MovieList.jsx
@@ -1,54 +1,10 @@
 import MovieCard from "./MovieCard";
-import movies from "../db/movies";
-import Button from "./Button";
-import { useState } from "react";
-
-// Props Version
+import { useMyContext } from "../Context/useMyContext";
 
 const MovieList = () => {
-  const [moviesArray, setMoviesArray] = useState(movies);
-
-  const copyMovies = [...movies];
-
-  const handleDateAscending = () => {
-    copyMovies.sort((a, b) => a.year - b.year);
-    setMoviesArray(copyMovies);
-  };
-
-  const handleDateDescending = () => {
-    copyMovies.sort((a, b) => b.year - a.year);
-    setMoviesArray(copyMovies);
-  };
-  const handleWorstRate = () => {
-    copyMovies.sort((a, b) => a.rate - b.rate);
-    setMoviesArray(copyMovies);
-  };
-
-  const handleBestRate = () => {
-    copyMovies.sort((a, b) => b.rate - a.rate);
-    setMoviesArray(copyMovies);
-  };
-  const handleAphabetical = () => {
-    copyMovies.sort((a, b) => a.title.localeCompare(b.title));
-    setMoviesArray(copyMovies);
-  };
-  const handleDescendingAlphabetical = () => {
-    copyMovies.sort((a, b) => b.title.localeCompare(a.title));
-    setMoviesArray(copyMovies);
-  };
-
+  const { moviesArray, setMoviesArray } = useMyContext();
   return (
     <>
-      <div className="flex items-center justify-center flex-wrap">
-        <Button action={handleBestRate} name="Rate ⬆" />
-        <Button action={handleWorstRate} name="Rate ⬇" />
-        <Button action={handleDateAscending} name="Date  ⬆" />
-        <Button action={handleDateDescending} name="Date  ⬇" />
-      </div>
-      <div className="flex items-center justify-center flex-wrap">
-        <Button action={handleAphabetical} name="A-Z" />
-        <Button action={handleDescendingAlphabetical} name="Z-A" />
-      </div>
       <article>
         {moviesArray.map((movie, index) => (
           <MovieCard moviesObject={movie} key={index} />
